:recycle: Refactor SocketBuf class
- Add StreamCleaner.hpp
